Move signal receiver over to signals.py
