+++ soft_deletion/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'soft_deletion.apps.SoftDeletionConfig'
