+++ soft_deletion/apps.py
@@ -1,3 +1,5 @@
+import sys
+
 from django.apps import AppConfig
 
 
@@ -5,3 +7,6 @@ class SoftDeletionConfig(AppConfig):
     name = 'soft_deletion'
     verbose_name = "Suppression"
 
+    def ready(self):
+        if 'migrate' not in sys.argv:
+            import soft_deletion.signals  # noqa
